// apps/web/src/shared/api/graphql/__generated__/schema.graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export interface Scalars {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Date: { input: any; output: any };
  DateTime: { input: any; output: any };
  JSON: { input: any; output: any };
  Time: { input: any; output: any };
  Upload: { input: any; output: any };
}
// ...
export interface ComponentEventEventFiltersInput {
  and?: InputMaybe<Array<InputMaybe<ComponentEventEventFiltersInput>>>;
  date?: InputMaybe<ComponentDateEventDateFiltersInput>;
  info?: InputMaybe<StringFilterInput>;
  location?: InputMaybe<ComponentLocationLocationFiltersInput>;
  not?: InputMaybe<ComponentEventEventFiltersInput>;
  or?: InputMaybe<Array<InputMaybe<ComponentEventEventFiltersInput>>>;
  schedule?: InputMaybe<ComponentEventScheduleFiltersInput>;
}
// ...
export interface ComponentLinksToMixesLink {
  __typename?: 'ComponentLinksToMixesLink';
  id: Scalars['ID']['output'];
  soundcloudLink?: Maybe<Scalars['String']['output']>;
  youtubeLink?: Maybe<Scalars['String']['output']>;
}

export interface ComponentLinksToMixesLinkFiltersInput {
  and?: InputMaybe<Array<InputMaybe<ComponentLinksToMixesLinkFiltersInput>>>;
  not?: InputMaybe<ComponentLinksToMixesLinkFiltersInput>;
  or?: InputMaybe<Array<InputMaybe<ComponentLinksToMixesLinkFiltersInput>>>;
  soundcloudLink?: InputMaybe<StringFilterInput>;
  youtubeLink?: InputMaybe<StringFilterInput>;
}
// ...
export interface ComponentLinksToSocialsLinksToSocialsFiltersInput {
  BCLink?: InputMaybe<StringFilterInput>;
  IGLink?: InputMaybe<StringFilterInput>;
  SCLink?: InputMaybe<StringFilterInput>;
  TGLink?: InputMaybe<StringFilterInput>;
  VKLink?: InputMaybe<StringFilterInput>;
  and?: InputMaybe<
    Array<InputMaybe<ComponentLinksToSocialsLinksToSocialsFiltersInput>>
  >;
  not?: InputMaybe<ComponentLinksToSocialsLinksToSocialsFiltersInput>;
  or?: InputMaybe<
    Array<InputMaybe<ComponentLinksToSocialsLinksToSocialsFiltersInput>>
  >;
}
// ...
export interface DateTimeFilterInput {
  and?: InputMaybe<Array<InputMaybe<Scalars['DateTime']['input']>>>;
  between?: InputMaybe<Array<InputMaybe<Scalars['DateTime']['input']>>>;
  contains?: InputMaybe<Scalars['DateTime']['input']>;
  containsi?: InputMaybe<Scalars['DateTime']['input']>;
  endsWith?: InputMaybe<Scalars['DateTime']['input']>;
  eq?: InputMaybe<Scalars['DateTime']['input']>;
  eqi?: InputMaybe<Scalars['DateTime']['input']>;
  gt?: InputMaybe<Scalars['DateTime']['input']>;
  gte?: InputMaybe<Scalars['DateTime']['input']>;
  in?: InputMaybe<Array<InputMaybe<Scalars['DateTime']['input']>>>;
  lt?: InputMaybe<Scalars['DateTime']['input']>;
  lte?: InputMaybe<Scalars['DateTime']['input']>;
  ne?: InputMaybe<Scalars['DateTime']['input']>;
  not?: InputMaybe<DateTimeFilterInput>;
  notContains?: InputMaybe<Scalars['DateTime']['input']>;
  notContainsi?: InputMaybe<Scalars['DateTime']['input']>;
  notIn?: InputMaybe<Array<InputMaybe<Scalars['DateTime']['input']>>>;
  notNull?: InputMaybe<Scalars['Boolean']['input']>;
  null?: InputMaybe<Scalars['Boolean']['input']>;
  or?: InputMaybe<Array<InputMaybe<Scalars['DateTime']['input']>>>;
  startsWith?: InputMaybe<Scalars['DateTime']['input']>;
}
// ...
export interface EventEntityResponse {
  __typename?: 'EventEntityResponse';
  data?: Maybe<EventEntity>;
}

export interface EventEntityResponseCollection {
  __typename?: 'EventEntityResponseCollection';
  data: Array<EventEntity>;
  meta: ResponseCollectionMeta;
}

export interface EventFiltersInput {
  and?: InputMaybe<Array<InputMaybe<EventFiltersInput>>>;
  createdAt?: InputMaybe<DateTimeFilterInput>;
  eventInfo?: InputMaybe<ComponentEventEventFiltersInput>;
  id?: InputMaybe<IdFilterInput>;
  mixes?: InputMaybe<MixFiltersInput>;
  name?: InputMaybe<StringFilterInput>;
  not?: InputMaybe<EventFiltersInput>;
  or?: InputMaybe<Array<InputMaybe<EventFiltersInput>>>;
  publishedAt?: InputMaybe<DateTimeFilterInput>;
  slug?: InputMaybe<StringFilterInput>;
  updatedAt?: InputMaybe<DateTimeFilterInput>;
}
// ...
export interface GenreFiltersInput {
  and?: InputMaybe<Array<InputMaybe<GenreFiltersInput>>>;
  createdAt?: InputMaybe<DateTimeFilterInput>;
  id?: InputMaybe<IdFilterInput>;
  mixes?: InputMaybe<MixFiltersInput>;
  name?: InputMaybe<StringFilterInput>;
  not?: InputMaybe<GenreFiltersInput>;
  or?: InputMaybe<Array<InputMaybe<GenreFiltersInput>>>;
  slug?: InputMaybe<StringFilterInput>;
  updatedAt?: InputMaybe<DateTimeFilterInput>;
}

export interface GenreInput {
  mixes?: InputMaybe<Array<InputMaybe<Scalars['ID']['input']>>>;
  name?: InputMaybe<Scalars['String']['input']>;
  slug?: InputMaybe<Scalars['String']['input']>;
}

export interface GenreRelationResponseCollection {
  __typename?: 'GenreRelationResponseCollection';
  data: Array<GenreEntity>;
}
// ...
export interface GuestFiltersInput {
  and?: InputMaybe<Array<InputMaybe<GuestFiltersInput>>>;
  createdAt?: InputMaybe<DateTimeFilterInput>;
  descriptionEn?: InputMaybe<StringFilterInput>;
  descriptionRu?: InputMaybe<StringFilterInput>;
  id?: InputMaybe<IdFilterInput>;
  mixes?: InputMaybe<MixFiltersInput>;
  name?: InputMaybe<StringFilterInput>;
  not?: InputMaybe<GuestFiltersInput>;
  or?: InputMaybe<Array<InputMaybe<GuestFiltersInput>>>;
  publishedAt?: InputMaybe<DateTimeFilterInput>;
  slug?: InputMaybe<StringFilterInput>;
  socials?: InputMaybe<ComponentLinksToSocialsLinksToSocialsFiltersInput>;
  updatedAt?: InputMaybe<DateTimeFilterInput>;
}
// ...
export interface GuestRelationResponseCollection {
  __typename?: 'GuestRelationResponseCollection';
  data: Array<GuestEntity>;
}
// ...
export interface IdFilterInput {
  and?: InputMaybe<Array<InputMaybe<Scalars['ID']['input']>>>;
  between?: InputMaybe<Array<InputMaybe<Scalars['ID']['input']>>>;
  contains?: InputMaybe<Scalars['ID']['input']>;
  containsi?: InputMaybe<Scalars['ID']['input']>;
  endsWith?: InputMaybe<Scalars['ID']['input']>;
  eq?: InputMaybe<Scalars['ID']['input']>;
  eqi?: InputMaybe<Scalars['ID']['input']>;
  gt?: InputMaybe<Scalars['ID']['input']>;
  gte?: InputMaybe<Scalars['ID']['input']>;
  in?: InputMaybe<Array<InputMaybe<Scalars['ID']['input']>>>;
  lt?: InputMaybe<Scalars['ID']['input']>;
  lte?: InputMaybe<Scalars['ID']['input']>;
  ne?: InputMaybe<Scalars['ID']['input']>;
  not?: InputMaybe<IdFilterInput>;
  notContains?: InputMaybe<Scalars['ID']['input']>;
  notContainsi?: InputMaybe<Scalars['ID']['input']>;
  notIn?: InputMaybe<Array<InputMaybe<Scalars['ID']['input']>>>;
  notNull?: InputMaybe<Scalars['Boolean']['input']>;
  null?: InputMaybe<Scalars['Boolean']['input']>;
  or?: InputMaybe<Array<InputMaybe<Scalars['ID']['input']>>>;
  startsWith?: InputMaybe<Scalars['ID']['input']>;
}
// ...
export interface Mix {
  __typename?: 'Mix';
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  date: Scalars['DateTime']['output'];
  descriptionEn?: Maybe<Scalars['String']['output']>;
  descriptionRu?: Maybe<Scalars['String']['output']>;
  event?: Maybe<EventEntityResponse>;
  genres?: Maybe<GenreRelationResponseCollection>;
  guests?: Maybe<GuestRelationResponseCollection>;
  image: UploadFileEntityResponse;
  linksToMixes?: Maybe<ComponentLinksToMixesLink>;
  locationLink?: Maybe<Scalars['String']['output']>;
  locationName?: Maybe<Scalars['String']['output']>;
  moods?: Maybe<MoodRelationResponseCollection>;
  name: Scalars['String']['output'];
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  show?: Maybe<ShowEntityResponse>;
  slug?: Maybe<Scalars['String']['output']>;
  tracklist?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
}
// ...
export interface MixEntity {
  __typename?: 'MixEntity';
  attributes?: Maybe<Mix>;
  id?: Maybe<Scalars['ID']['output']>;
}

export interface MixEntityResponse {
  __typename?: 'MixEntityResponse';
  data?: Maybe<MixEntity>;
}

export interface MixEntityResponseCollection {
  __typename?: 'MixEntityResponseCollection';
  data: Array<MixEntity>;
  meta: ResponseCollectionMeta;
}

export interface MixFiltersInput {
  and?: InputMaybe<Array<InputMaybe<MixFiltersInput>>>;
  createdAt?: InputMaybe<DateTimeFilterInput>;
  date?: InputMaybe<DateTimeFilterInput>;
  descriptionEn?: InputMaybe<StringFilterInput>;
  descriptionRu?: InputMaybe<StringFilterInput>;
  event?: InputMaybe<EventFiltersInput>;
  genres?: InputMaybe<GenreFiltersInput>;
  guests?: InputMaybe<GuestFiltersInput>;
  id?: InputMaybe<IdFilterInput>;
  linksToMixes?: InputMaybe<ComponentLinksToMixesLinkFiltersInput>;
  locationLink?: InputMaybe<StringFilterInput>;
  locationName?: InputMaybe<StringFilterInput>;
  moods?: InputMaybe<MoodFiltersInput>;
  name?: InputMaybe<StringFilterInput>;
  not?: InputMaybe<MixFiltersInput>;
  or?: InputMaybe<Array<InputMaybe<MixFiltersInput>>>;
  publishedAt?: InputMaybe<DateTimeFilterInput>;
  show?: InputMaybe<ShowFiltersInput>;
  slug?: InputMaybe<StringFilterInput>;
  tracklist?: InputMaybe<StringFilterInput>;
  updatedAt?: InputMaybe<DateTimeFilterInput>;
}
// ...
export interface MoodFiltersInput {
  and?: InputMaybe<Array<InputMaybe<MoodFiltersInput>>>;
  createdAt?: InputMaybe<DateTimeFilterInput>;
  description?: InputMaybe<StringFilterInput>;
  id?: InputMaybe<IdFilterInput>;
  mixes?: InputMaybe<MixFiltersInput>;
  name?: InputMaybe<StringFilterInput>;
  not?: InputMaybe<MoodFiltersInput>;
  or?: InputMaybe<Array<InputMaybe<MoodFiltersInput>>>;
  slug?: InputMaybe<StringFilterInput>;
  updatedAt?: InputMaybe<DateTimeFilterInput>;
}

export interface MoodInput {
  description?: InputMaybe<Scalars['String']['input']>;
  mixes?: InputMaybe<Array<InputMaybe<Scalars['ID']['input']>>>;
  name?: InputMaybe<Scalars['String']['input']>;
  slug?: InputMaybe<Scalars['String']['input']>;
}

export interface MoodRelationResponseCollection {
  __typename?: 'MoodRelationResponseCollection';
  data: Array<MoodEntity>;
}
// ...
export interface Pagination {
  __typename?: 'Pagination';
  page: Scalars['Int']['output'];
  pageCount: Scalars['Int']['output'];
  pageSize: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
}
// ...
export interface ResponseCollectionMeta {
  __typename?: 'ResponseCollectionMeta';
  pagination: Pagination;
}
// ...
export interface ShowEntityResponse {
  __typename?: 'ShowEntityResponse';
  data?: Maybe<ShowEntity>;
}

export interface ShowEntityResponseCollection {
  __typename?: 'ShowEntityResponseCollection';
  data: Array<ShowEntity>;
  meta: ResponseCollectionMeta;
}

export interface ShowFiltersInput {
  and?: InputMaybe<Array<InputMaybe<ShowFiltersInput>>>;
  createdAt?: InputMaybe<DateTimeFilterInput>;
  descriptionEn?: InputMaybe<StringFilterInput>;
  descriptionRu?: InputMaybe<StringFilterInput>;
  id?: InputMaybe<IdFilterInput>;
  mixes?: InputMaybe<MixFiltersInput>;
  name?: InputMaybe<StringFilterInput>;
  not?: InputMaybe<ShowFiltersInput>;
  or?: InputMaybe<Array<InputMaybe<ShowFiltersInput>>>;
  publishedAt?: InputMaybe<DateTimeFilterInput>;
  slug?: InputMaybe<StringFilterInput>;
  socials?: InputMaybe<ComponentLinksToSocialsLinksToSocialsFiltersInput>;
  updatedAt?: InputMaybe<DateTimeFilterInput>;
  youtubePlayerLink?: InputMaybe<StringFilterInput>;
}
// ...
export interface StringFilterInput {
  and?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
  between?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
  contains?: InputMaybe<Scalars['String']['input']>;
  containsi?: InputMaybe<Scalars['String']['input']>;
  endsWith?: InputMaybe<Scalars['String']['input']>;
  eq?: InputMaybe<Scalars['String']['input']>;
  eqi?: InputMaybe<Scalars['String']['input']>;
  gt?: InputMaybe<Scalars['String']['input']>;
  gte?: InputMaybe<Scalars['String']['input']>;
  in?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
  lt?: InputMaybe<Scalars['String']['input']>;
  lte?: InputMaybe<Scalars['String']['input']>;
  ne?: InputMaybe<Scalars['String']['input']>;
  not?: InputMaybe<StringFilterInput>;
  notContains?: InputMaybe<Scalars['String']['input']>;
  notContainsi?: InputMaybe<Scalars['String']['input']>;
  notIn?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
  notNull?: InputMaybe<Scalars['Boolean']['input']>;
  null?: InputMaybe<Scalars['Boolean']['input']>;
  or?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
  startsWith?: InputMaybe<Scalars['String']['input']>;
}
// ...
export interface UploadFileEntityResponse {
  __typename?: 'UploadFileEntityResponse';
  data?: Maybe<UploadFileEntity>;
}